Improve preview rendering precision [1/2]
The editor calls Math.floor() on the pixel count for cell sizes,
so it may leave a transparent edge at the right/bottom.  When
drawing the entire editor to the preview, the transparency would
be included in the copy.

This had been especially visible when experimenting with the size
of the editor; at 360px with 11 cells, only 352px were drawn,
leaving over 2px between blocks in the preview after scale-down.