Clarify (de-unify) resizing.  Known issues.
Expand, paint, shrink, re-expand: the painting is gone.  This happens
because we're not updating the savedCells with the current cells.

Some of the updateEditor optimization is gone.  E.g. if a sub-cell is
painted with the same color, the editor and preview are probably re-
painted without visible effect.  The root cause is a problem with the
sequence numbers; the optimization was removed because the initial
paint wasn't happening, because block's drawCount=0 and editorState=0.
editorState=-1 fixed that, but broke all updates, as it now equaled
the view.editorState -1.  We should initialize as undefined and treat
it specially, instead of playing with sentinels like -1/0/1 or adding
-2 to the mix.