TypeScript 5 and Firefox 52, at last
This finally brings modern code back to browsers of 2017, and operating
systems of 2007.

I do not fully understand this.  I don't think there are polyfills in
play, but trying to add them (babel and core-js) makes it stop loading
in my Firefox 52 VM.  Bringing in the node-resolve plugin doesn't help:
require is still not defined.

Instead, I ask TypeScript to generate ES2015 and hope for the best in my
2017 browser target.