Update development instructions, preserve `dist`
I actually did development with `php -S` and completely removing dist
would leave it stranded in the (deleted) directory.  Solve that for
other people by not deleting the whole thing unconditionally.

Also, write directions for webpack-dev-server, which actually ships as a
dependency in package-lock.json, so that other people don't *also* need
PHP "or something" to serve the directory.

Finally, add `--frozen-lockfile` to the `yarn install` so it does
*precisely* what I want, which is to be like `npm ci` or `composer
install`.