Fix the guides
We shouldn't set the block dirty.  That causes it to redraw all cells,
even if only the guide color changed.  Instead, let it copy its internal
canvas back out to the editor canvas.

Since updateEditor() knows when it's copying a block, have it take all
responsibility for drawing the guides on afterward.  As such, drawGuides
is never called with an empty context, so it doesn't have to check for
that, nor create one.

The glitch in positioning the guide lines comes about because it was
written with 0.5-px adjustment, while nothing else is.