Post-tabs print view enhancements
Maybe there should be a 'print' button that makes a nice preview, but
until then... there's this.  Portrait layout lets the large preview
print on one page instead of three, and hiding the mini preview when
printing the block uses one page instead of two.

It really is the best lemonade we can hastily make with our lemons.