Let `drawPreviewBlocks()` scale its own sources
Now that RenderData has the Quilt in it, we can support multiple blocks
on the preview by letting drawPreviewBlocks() read the Quilt.

Quilt gains a list of block indices (currently all 0; there is only one
block in the UI) for drawPreviewBlocks() to read out.  This moves more
pieces into place to have multiple blocks and "block drawing tools" on
the Preview tab.