Inline `CELL_QUADRANTS` lookup
Directly generate a number, instead of constructing a string and doing a
lookup in an object with it.

Inline the "selected color" constant as well.  Pass the lookup result
directly to paintSubCell.

Adjust the bounds on paintSubCell so that it doesn't have to check j.
Update it to reject negative values of i or color.