Unify the mini/preview pathways
New class in view.ts, Previewer, which is responsible for rendering "a
preview" onto a configured canvas, resizing if necessary to fit the
current available space.

This means that we no longer need separate updatePreview and
updateMiniPreview functions; they've become Previewer#render(), and it's
simply a matter of picking the `previewer` or `miniPreviewer` variable
to render.

It also means that the ViewState (view and miniView) have disappeared
from ui.ts, as they are private to the Previewer.  Since it _exists,_ it
makes a good place to hold this "last drawn" state.

The `[MINI]_PREVIEW_DRAW_{WIDTH|HEIGHT}` variables also disappear, as
that information is the `drawSize:Rect` getter of Previewer instances.