Add Quilt.resize() method
Make Quilt.blockMap a private, 2D array.  Like the cells in a block,
remember what was on the quilt if sized down and back up: Quilt.shape
now specifies the accessible window over the possibly-larger blockMap
data.

To give the preview drawing code access again, provide an iterator over
the rows, that yields an iterable over the columns.